+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "export": "next export"
+    "export": "next export",
+    "test": "jest"
   },
   "dependencies": {
     "next": "^11.0.1",
@@ -16,10 +17,15 @@
     "styled-components": "^5.3.0"
   },
   "devDependencies": {
+    "@types/jest": "^26.0.24",
     "@types/node": "^16.4.3",
     "@types/react": "^17.0.15",
     "@types/styled-components": "^5.1.11",
     "babel-plugin-styled-components": "^1.13.2",
+    "ts-jest": "^27.0.4",
     "typescript": "^4.3.5"
+  },
+  "peerDependencies": {
+    "jest": "^27.0.6"
   }
 }
