+++ src/components/CleanData/cleanData.test.ts
@@ -1,0 +1,5 @@
+import { cleanData } from "./cleanData"
+
+it("hoge", () => {
+  expect(cleanData("hoge")).toBe("fuga")
+})
