+++ src/components/CleanData/index.tsx
@@ -1,22 +1,11 @@
 import { css } from "styled-components"
 import { Textarea } from "../Textarea"
 import { useInput } from "../../hooks/useInput"
-
-export const clean = (text: string): string => {
-  const converter = (line: string): string => {
-    let result = line
-    result = result.replace(/[^\d\t]/g, "")
-    return result
-  }
-
-  return text.split("\n")
-    .map(converter)
-    .join("\n")
-}
+import { cleanData } from "./cleanData"
 
 export function CleanData(): JSX.Element {
   const { text, change } = useInput()
-  const cleaned = clean(text)
+  const cleaned = cleanData(text)
 
   return (
     <div css={css`
